refactor: util 모듈 android, jvm 의존성에 따라 분리
