+++ app/build.gradle.kts
@@ -57,7 +57,8 @@ dependencies {
     implementation(projects.domain)
     implementation(projects.data)
     implementation(projects.core.ui)
-    implementation(projects.core.util)
+    implementation(projects.core.util.android)
+    implementation(projects.core.util.jvm)
     implementation(projects.core.designsystem)
     implementation(projects.core.navigation)
