+++ build-logic/convention/src/main/kotlin/ModuleDomainConventionPlugin.kt
@@ -8,6 +8,8 @@ class ModuleDomainConventionPlugin : BaseConventionPlugin({
     }
 
     dependencies {
+        implementation(project(":core:util:jvm"))
+
         implementation(libs.javax.inject)
     }
 })
