+++ core/ui/build.gradle.kts
@@ -9,6 +9,7 @@ android {
 }
 
 dependencies {
-    implementation(projects.core.util)
+    implementation(projects.core.util.android)
+    implementation(projects.core.util.jvm)
     implementation(projects.core.designsystem)
 }
