+++ core/util/android/.gitignore
