+++ core/util/android/src/main/AndroidManifest.xml
