+++ core/util/jvm/build.gradle.kts
@@ -1,0 +1,9 @@
+import com.teamyg.parfait.buildlogic.utils.extensions.implementation
+
+plugins {
+    alias(libs.plugins.parfait.kotlin.jvm)
+}
+
+dependencies {
+    implementation(libs.kermit)
+}
